BackTracker made optional in SimPhotonCounter module.
SimPhotonCounter was recently changed to use BackTracker.
Backtracker was only used for a sanity check that would result in warnings
being printed. Now if BackTracker service is not configured, a non-fatal
error message is printed, and the check is skipped.
This makes the change a non-breaking one even when the existing jobs do
not set up BackTracker service.